refactor(nlp): remove orphaned NLP backend scaffolding
The Phase 2.1-2.4 NLPBackendRegistry / MeCabBackend / SpaCyBackend /
NLPHttpAdapter scaffolding was never wired to any call site. The
actual runtime path is:

  renderer getBackend().tokenize()
    -> HTTP /nlp/tokenize
    -> plugin_registry.get_active().LANGUAGE_TOKENIZE

and language switches restart the Python backend with a new active
module (single-language-per-process). The abstraction layer was
speculative and orphaned.

Deleted:
- src/shared/nlp-backend-abstraction.ts
- src/shared/nlp-backend-registry.{ts,test.ts}
- src/shared/nlp-backend-factory.{ts,test.ts}
- src/shared/mecab-backend.{ts,test.ts}
- src/shared/spacy-backend.{ts,test.ts}
- src/shared/nlp-http-adapter.{ts,test.ts}
- src/renderer/hooks/useNLPTokenizer.integration.test.ts
- src/renderer/hooks/useNLPTokenizer.real-backend.test.ts
- src/renderer/components/test/TokenizerTestComponent.{tsx,css}
- src/root-of-app/routes/nlp_backends.py

Updated:
- useNLPTokenizer.ts: inline minimal MorphToken/TokenizationResult
  types; route tokenization through getBackend() like the rest of
  the app. No LanguageContext coupling.
- useNLPTokenizer.test.ts: import types from the hook itself.
- renderer/hooks/index.ts: drop useNLPTokenizer barrel export
  (unused).
- shared/index.ts: drop re-exports that pointed at deleted modules.
- server.py / nlp.py: drop nlp_backends router and the dead
  per-request language parameter (backend is single-language).
- vitest.config.ts: drop includes for the deleted test files.

Typecheck PASS. Vitest: 3422 passed / 1 pre-existing unrelated fail.
