+++ src/renderer/hooks/index.ts
@@ -19,7 +19,6 @@ export {
   warmTranslationCache,
   getCachedTranslation
 } from './useTranslation';
-export { useNLPTokenizer, getCachedNLPTokenization, clearNLPTokenizationCache } from './useNLPTokenizer';
 export { useWordHover, getGlobalHoverManager } from './useWordHover';
 export { useMediaStats } from './useMediaStats';
