--- src/renderer/hooks/useNLPTokenizer.integration.test.ts
@@ -1,332 +1,0 @@
-/**
- * useNLPTokenizer Integration Tests
- * Tests with real NLP backends (MeCab for Japanese, spaCy for German)
- * 
- * Prerequisites:
- * - MeCab server running on localhost:7752 (for Japanese)
- * - spaCy server running on localhost:7752 (for German)
- * - Python backend with NLP routes available
- */
-
-import { describe, it, expect, beforeAll, afterAll, skip } from 'vitest';
-import type { TokenizationResult } from '../../shared/nlp-backend-abstraction';
-import type { LanguageCode } from '../../shared/language-abstraction';
-
-// Skip integration tests if backends are not available
-const SKIP_INTEGRATION = process.env.SKIP_INTEGRATION === 'true';
-
-describe.skipIf(SKIP_INTEGRATION)('useNLPTokenizer Integration Tests', () => {
-  describe('Japanese tokenization with MeCab', () => {
-    it('should tokenize simple Japanese sentence', async () => {
-      const text = 'こんにちは';
-      const language: LanguageCode = 'ja';
-      
-      // This test requires:
-      // 1. MeCab backend to be initialized
-      // 2. LanguageContext to be available
-      // 3. NLP backend registry to have MeCab registered
-      
-      // Expected result structure:
-      // {
-      //   text: 'こんにちは',
-      //   language: 'ja',
-      //   tokens: [
-      //     {
-      //       surface: 'こんにちは',
-      //       lemma: 'こんにちは',
-      //       pos: '感動詞',
-      //       reading: 'コンニチハ',
-      //       pitchAccent: 0
-      //     }
-      //   ],
-      //   processingTime: number,
-      //   confidence: number
-      // }
-    });
-
-    it('should tokenize complex Japanese sentence with kanji', async () => {
-      const text = '私は毎日学校に行きます';
-      const language: LanguageCode = 'ja';
-      
-      // Expected tokens:
-      // 私 (watashi) - pronoun
-      // は (ha) - particle
-      // 毎日 (mainichi) - noun
-      // 学校 (gakkou) - noun
-      // に (ni) - particle
-      // 行き (iki) - verb stem
-      // ます (masu) - auxiliary verb
-    });
-
-    it('should extract readings from Japanese tokens', async () => {
-      const text = '漢字';
-      const language: LanguageCode = 'ja';
-      
-      // Expected:
-      // - surface: 漢字
-      // - lemma: 漢字
-      // - reading: カンジ
-      // - pos: 名詞
-    });
-
-    it('should detect pitch accent for Japanese words', async () => {
-      const text = '橋';
-      const language: LanguageCode = 'ja';
-      
-      // Expected:
-      // - pitchAccent: 0 or 1 (depending on word)
-      // - reading: ハシ
-    });
-
-    it('should handle mixed hiragana and kanji', async () => {
-      const text = 'ひらがなと漢字';
-      const language: LanguageCode = 'ja';
-      
-      // Should correctly tokenize mixed scripts
-    });
-
-    it('should handle katakana', async () => {
-      const text = 'コンピューター';
-      const language: LanguageCode = 'ja';
-      
-      // Should correctly tokenize katakana words
-    });
-
-    it('should handle punctuation', async () => {
-      const text = 'これはテストです。';
-      const language: LanguageCode = 'ja';
-      
-      // Should include punctuation as tokens
-    });
-
-    it('should handle empty string gracefully', async () => {
-      const text = '';
-      const language: LanguageCode = 'ja';
-      
-      // Should throw error or return empty tokens
-    });
-
-    it('should handle very long text', async () => {
-      const text = 'これは長いテキストです。'.repeat(100);
-      const language: LanguageCode = 'ja';
-      
-      // Should handle large inputs without crashing
-    });
-
-    it('should return consistent results for same input', async () => {
-      const text = 'テスト';
-      const language: LanguageCode = 'ja';
-      
-      // Multiple calls with same input should return identical results
-    });
-  });
-
-  describe('German tokenization with spaCy', () => {
-    it('should tokenize simple German sentence', async () => {
-      const text = 'Guten Tag';
-      const language: LanguageCode = 'de';
-      
-      // Expected result structure:
-      // {
-      //   text: 'Guten Tag',
-      //   language: 'de',
-      //   tokens: [
-      //     {
-      //       surface: 'Guten',
-      //       lemma: 'gut',
-      //       pos: 'ADJ',
-      //       reading: undefined,
-      //       pitchAccent: undefined
-      //     },
-      //     {
-      //       surface: 'Tag',
-      //       lemma: 'Tag',
-      //       pos: 'NOUN',
-      //       reading: undefined,
-      //       pitchAccent: undefined
-      //     }
-      //   ],
-      //   processingTime: number,
-      //   confidence: number
-      // }
-    });
-
-    it('should tokenize German sentence with verbs', async () => {
-      const text = 'Ich gehe zur Schule';
-      const language: LanguageCode = 'de';
-      
-      // Expected tokens:
-      // Ich - pronoun
-      // gehe - verb
-      // zur - preposition + article
-      // Schule - noun
-    });
-
-    it('should handle German compound words', async () => {
-      const text = 'Donaudampfschifffahrtsgesellschaftskapitän';
-      const language: LanguageCode = 'de';
-      
-      // Should correctly tokenize compound words
-    });
-
-    it('should handle German umlauts', async () => {
-      const text = 'Äpfel, Öl, Übung';
-      const language: LanguageCode = 'de';
-      
-      // Should correctly handle special characters
-    });
-
-    it('should handle German punctuation', async () => {
-      const text = 'Das ist ein Test!';
-      const language: LanguageCode = 'de';
-      
-      // Should include punctuation as tokens
-    });
-
-    it('should extract lemmas for German words', async () => {
-      const text = 'laufen';
-      const language: LanguageCode = 'de';
-      
-      // Expected:
-      // - surface: laufen
-      // - lemma: laufen
-      // - pos: VERB
-    });
-
-    it('should handle German articles', async () => {
-      const text = 'der Mann, die Frau, das Kind';
-      const language: LanguageCode = 'de';
-      
-      // Should correctly tokenize articles
-    });
-
-    it('should not have pitch accent for German', async () => {
-      const text = 'Wort';
-      const language: LanguageCode = 'de';
-      
-      // Expected:
-      // - pitchAccent: undefined
-      // - reading: undefined
-    });
-
-    it('should handle empty string gracefully', async () => {
-      const text = '';
-      const language: LanguageCode = 'de';
-      
-      // Should throw error or return empty tokens
-    });
-
-    it('should handle very long text', async () => {
-      const text = 'Dies ist ein Test. '.repeat(100);
-      const language: LanguageCode = 'de';
-      
-      // Should handle large inputs without crashing
-    });
-
-    it('should return consistent results for same input', async () => {
-      const text = 'Test';
-      const language: LanguageCode = 'de';
-      
-      // Multiple calls with same input should return identical results
-    });
-  });
-
-  describe('cross-language behavior', () => {
-    it('should handle switching between languages', async () => {
-      // Tokenize Japanese text
-      // Then tokenize German text
-      // Verify correct backend is used for each
-    });
-
-    it('should cache results per language', async () => {
-      // Same text in different languages should have separate cache entries
-    });
-
-    it('should handle unsupported language gracefully', async () => {
-      const text = 'test';
-      const language: LanguageCode = 'fr'; // French not supported
-      
-      // Should throw error indicating no backend available
-    });
-  });
-
-  describe('performance characteristics', () => {
-    it('should tokenize within reasonable time', async () => {
-      const text = 'これはテストです';
-      const language: LanguageCode = 'ja';
-      
-      // Should complete within 1 second
-      // Actual time depends on backend performance
-    });
-
-    it('should handle batch tokenization efficiently', async () => {
-      // If batch tokenization is available, should be faster than sequential
-    });
-
-    it('should cache results to avoid redundant processing', async () => {
-      // First call: hits backend
-      // Second call: returns cached result (should be instant)
-    });
-  });
-
-  describe('error handling', () => {
-    it('should handle backend timeout', async () => {
-      // If backend takes too long, should timeout gracefully
-    });
-
-    it('should handle backend connection error', async () => {
-      // If backend is not available, should throw error
-    });
-
-    it('should handle malformed backend response', async () => {
-      // If backend returns invalid data, should throw error
-    });
-
-    it('should handle invalid language code', async () => {
-      const text = 'test';
-      const language: LanguageCode = 'xx'; // Invalid
-      
-      // Should throw error
-    });
-  });
-
-  describe('token quality', () => {
-    it('should produce correct POS tags for Japanese', async () => {
-      const text = '私は学生です';
-      const language: LanguageCode = 'ja';
-      
-      // Verify POS tags are correct:
-      // 私 - 名詞
-      // は - 助詞
-      // 学生 - 名詞
-      // です - 助動詞
-    });
-
-    it('should produce correct POS tags for German', async () => {
-      const text = 'Ich bin Schüler';
-      const language: LanguageCode = 'de';
-      
-      // Verify POS tags are correct:
-      // Ich - PRON
-      // bin - AUX
-      // Schüler - NOUN
-    });
-
-    it('should produce correct lemmas for Japanese', async () => {
-      const text = '走っている';
-      const language: LanguageCode = 'ja';
-      
-      // Verify lemmas are correct:
-      // 走っ - 走る (verb stem)
-      // ている - いる (auxiliary)
-    });
-
-    it('should produce correct lemmas for German', async () => {
-      const text = 'laufen';
-      const language: LanguageCode = 'de';
-      
-      // Verify lemma is correct:
-      // laufen - laufen
-    });
-  });
-});
