+++ src/root-of-app/server.py
@@ -29,7 +29,6 @@ _log(f"::QUIT_TOKEN::{config.QUIT_TOKEN}")
 
 # ── Route modules ──
 from routes import anki, nlp, ocr, llm, voice
-from routes import nlp_backends
 
 # ── FastAPI app ──
 app = FastAPI()
@@ -48,7 +47,6 @@ app.include_router(nlp.router)
 app.include_router(ocr.router)
 app.include_router(llm.router)
 app.include_router(voice.router)
-app.include_router(nlp_backends.router)
 
 
 # ── Middleware ──
