+++ vitest.config.ts
@@ -58,11 +58,6 @@ export default defineConfig({
             'src/shared/bridges/**/*.test.ts',
             'src/shared/backends/**/*.test.ts',
             'src/shared/platform.test.ts',
-            'src/shared/nlp-backend-registry.test.ts',
-            'src/shared/nlp-backend-factory.test.ts',
-            'src/shared/mecab-backend.test.ts',
-            'src/shared/spacy-backend.test.ts',
-            'src/shared/nlp-http-adapter.test.ts',
           ],
         },
       },
